+++ src/main/java/com/example/prueba/service/IService/ICuentaService.java
@@ -5,8 +5,8 @@ import com.example.prueba.Entity.Cuenta;
 import java.util.Optional;
 
 public interface ICuentaService {
-    Cuenta crearCuenta(int id, String tipoCuenta, long GMF);
-    void actualizarSaldo(int id, double nuevoSaldo);
-    void cambiarEstado(int id, String nuevoEstado);
-    Optional<Cuenta> obtenerCuentaPorId(int id);
+    Cuenta crearCuenta(long id, String tipoCuenta, long GMF);
+    void actualizarSaldo(long id, double nuevoSaldo);
+    void cambiarEstado(long id, String nuevoEstado);
+    Optional<Cuenta> obtenerCuentaPorId(long id);
 }
